31/08/2021-6

+++ src/Interfaces/Guitar.java
@@ -15,7 +15,7 @@ public class Guitar implements Instrument{
 
  //getter
 
-    public Integer getParameter() {
+    public double getParameter() {
         return stringsNumber;
     }
